+++ NEWS.md
@@ -2,6 +2,7 @@ The latest release of the MSEtool package is available on [CRAN](https://CRAN.R-
 
 ## MSEtool 1.2.3
 - Users can choose to use `SRA_scope` while conditioning on either catch or effort. Function will also return a markdown report of model fits.
+- Vignette links are now available through the `?MSEtool` help page.
 
 ## MSEtool 1.2.2
 - `SS2OM` now has an option for selecting male or female life history parameters.